not fix: the FAT

+++ src/portio.c
@@ -33,7 +33,7 @@ void out16(uint16_t port, uint16_t data) {
 }
 
 uint16_t in16(uint16_t port) {
-    uint8_t result;
+    uint16_t result;
     __asm__ volatile(
         "inb %1, %0" 
         : "=a"(result)
